Move test from bin/dmtest.rs to a unit test in dm.rs.
This will help with automation.
