+++ Cargo.toml
@@ -10,10 +10,6 @@ readme = "README.md"
 keywords = ["Linux", "device", "mapper", "libdm", "storage"]
 license = "MPL-2.0"
 
-[[bin]]
-name = "dmtest"
-path = "src/bin/dmtest.rs"
-
 [dependencies]
 libc = "0.2"
 nix = "0.5"
